+++ ci_environment/google-chrome/metadata.rb
@@ -1,0 +1,7 @@
+name             'google-chrome'
+maintainer       'Travis CI GmbH'
+maintainer_email 'contact+travis-cookbooks-google-chrome@travis-ci.com'
+license          'MIT LISENCE'
+description      'Installs/Configures google-chrome'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'